refactor spatial subdivisions
