+++ sheerwater/spatial_subdivisions/__init__.py
@@ -1,0 +1,22 @@
+"""Spatial subdivision modulel."""
+
+from .utils import (masks_to_polygons, clip_region, clip_by_geometry,
+                    apply_mask, clip_with_mask, clip_station_grid,
+                    nonuniform_grid)
+from .spatial_subdivisions import (clean_spatial_subdivision_name, get_spatial_subdivision_level,
+                                  polygon_subdivision_geodataframe, polygon_subdivision_labels,
+                                  space_grouping_labels, reconcile_country_name)
+
+__all__ = ["masks_to_polygons",
+           "clip_region",
+           "clip_by_geometry",
+           "apply_mask",
+           "clip_with_mask",
+           "clip_station_grid",
+           "clean_spatial_subdivision_name",
+           "get_spatial_subdivision_level",
+           "polygon_subdivision_geodataframe",
+           "polygon_subdivision_labels",
+           "space_grouping_labels",
+           "reconcile_country_name",
+           "nonuniform_grid"]
